# Inline singleton splats