--- a/base/compiler/utilities.jl
+++ b/base/compiler/utilities.jl
@@ -133,6 +133,9 @@ function specialize_method(method::Method, @nospecialize(atypes), sparams::Simpl
     return ccall(:jl_specializations_get_linfo, Ref{MethodInstance}, (Any, Any, Any), method, atypes, sparams)
 end
 
+specialize_method(lookup::SimpleVector, preexisting::Bool=false) =
+    specialize_method(lookup[3], lookup[1], lookup[2], preexisting)
+
 # This function is used for computing alternate limit heuristics
 function method_for_inference_heuristics(method::Method, @nospecialize(sig), sparams::SimpleVector)
     if isdefined(method, :generator) && method.generator.expand_early && may_invoke_generator(method, sig, sparams)