@martinholters 2020-06-08
I know my way around the optimizer too little to review in detail, but IIUC `analyze_type_for_apply_rewrite` duplicates a bit of work already done in `abstract_iteration` during inference. Is that just due to a lack of infrastructure to pipe the information from the latter to the former, or is there some general need to re-establish the types obtained during iteration? (Not that it matters much now, but if we want to get fancier with the inliner in the future, it would be unfortunate to duplicate more of that logic.)