@martinholters 2020-06-09
Thanks for confirming. (And by all means, that does not mean I disapprove of this PR in its current form in any way.)